refactor: extract AccountManager into AppState extension (Stage 4a)
Stage 4a — penultimate sub-extraction in the AppState god-object split.
Moves all account lifecycle + bunker-URI + persistence logic out of
AppState.swift into a dedicated extension file. Cleanup follow-up to
the petname-removal PR (#42) that simplified the surface first.

Moves out of Clave/AppState.swift (16 methods + 1 MARK header):
- var bunkerURI (computed) + bunkerURI(for:) (per-account)
- loadAccounts (was private, now internal — called by loadState)
- recoverAccountsFromKeychainIfNeeded (Stage 2 deferred — moves now)
- cleanupOrphanLegacyKeychainEntry (Stage 2 deferred — moves now)
- persistAccountsList (private, stays private — only inter-extension)
- persistAccounts
- persistCurrentAccountPubkey (private, stays private)
- switchToAccount(pubkey:)
- addAccount(nsec:)
- generateAccount()
- deleteAccount(pubkey:)
- rotateBunkerSecret()
- importKey / generateKey / deleteKey (legacy wrappers preserved)
- refreshBunkerSecret() (no-op kept for source-compat)
- "// MARK: - Multi-account methods (Task 5)" header

Into the new file:
- Clave/AppState+AccountManager.swift (312 LOC) — extension AppState

What stays in AppState.swift:
- @Observable stored properties (accounts, currentAccount,
  bunkerSecretsTick, etc.) — Swift forbids stored properties in
  extensions. Pointer comment added at the multi-account state block.
- init() with NotificationCenter observers
- loadState() coordinator (calls into the extension via cross-file
  extension dispatch — works because all are extension AppState on
  the same final class in the same module)

One AppState.swift access change required:
- bunkerSecretsTick was `private(set) var` (file-scope setter).
  rotateBunkerSecret in the extension needs to mutate it across
  files, so dropped `private(set)` to internal-set. Same constraint
  pattern as Stage 3b's persistAccounts widen.

Cross-extension calls all resolve correctly:
- addAccount calls registerWithProxy (Stage 3c) + fetchProfileIfNeeded
  (Stage 3b) — works
- deleteAccount calls unregisterWithProxy + unpairClientWithProxy
  (Stage 3c) + cachedImageURL (Stage 3b) — works
- switchToAccount calls loadCachedProfileImage + fetchProfileIfNeeded
  (Stage 3b) — works

Zero behavior change:
- Function bodies preserved byte-for-byte
- All `private` modifiers preserved on persistAccountsList +
  persistCurrentAccountPubkey (only called within extension file —
  same-file private access)
- Modifiers loosened only where strictly required by cross-file
  calls (loadAccounts, recoverAccountsFromKeychainIfNeeded,
  cleanupOrphanLegacyKeychainEntry: private -> internal because
  loadState in AppState.swift calls them)

Test impact: zero changes.
- AppStateMultiAccountTests calls appState.<method> via extension
  dispatch — works unchanged.

External callers unchanged:
- ClaveApp + 7 view files call appState.<method> — extension dispatch.

No pbxproj edits — Clave/ directory is auto-synced.

Verification:
- xcodebuild test -skip-testing:ClaveUITests on iPhone 17 / iOS 26.4:
  - Pre-baseline (main @ f51bc18): 231 passed / 0 failed
  - Post-extraction (this branch): 231 passed / 0 failed
  - ** TEST SUCCEEDED ** in both runs

AppState.swift: 893 -> 611 LOC (-282, -31.6%).
Combined Stages 1+2+3+petname+4a: 2,338 -> 611 LOC (-73.9% from sprint
start).

Stage 4b (PendingApprovalCoordinator, ~290 LOC) is the final
extraction. After it lands, AppState should reach the BACKLOG
sketched ~300 LOC slim target.
